+++ main.go
@@ -6,6 +6,8 @@ import (
 	"os"
 
 	"github.com/gin-gonic/gin"
+	"github.com/iver-wharf/wharf-core/pkg/ginutil"
+	"github.com/iver-wharf/wharf-core/pkg/logger"
 	"github.com/iver-wharf/wharf-provider-github/docs"
 	"github.com/iver-wharf/wharf-provider-github/internal/httputils"
 
@@ -33,6 +35,8 @@ type importBody struct {
 
 const buildDefinitionFileName = ".wharf-ci.yml"
 
+var log = logger.NewScoped("WHARF-PROVIDER-GITHUB")
+
 // @title Wharf provider API for GitHub
 // @description Wharf backend API for integrating GitHub repositories with
 // @description the Wharf main API.
@@ -48,7 +52,7 @@ func main() {
 		err    error
 	)
 	if err = loadEmbeddedVersionFile(); err != nil {
-		fmt.Println("Failed to read embedded version.yaml file:", err)
+		log.Error().WithError(err).Message("Failed to read embedded version.yaml.")
 		os.Exit(1)
 	}
 	if config, err = loadConfig(); err != nil {
@@ -62,15 +66,23 @@ func main() {
 		client, err := httputils.NewClientWithCerts(config.CA.CertsFile)
 		if err != nil {
 			fmt.Println("Failed to get net/http.Client with certs:", err)
+			log.Error().WithError(err).Message("Failed to get net/http.Client with certs.")
 			os.Exit(1)
 		}
 		http.DefaultClient = client
 	}
 
-	r := gin.Default()
+	gin.DefaultWriter = ginutil.DefaultLoggerWriter
+	gin.DefaultErrorWriter = ginutil.DefaultLoggerWriter
+
+	r := gin.New()
+	r.Use(
+		ginutil.DefaultLoggerHandler,
+		ginutil.RecoverProblem,
+	)
 
 	if config.HTTP.CORS.AllowAllOrigins {
-		fmt.Printf("Allowing CORS\n")
+		log.Info().Message("Allowing all origins in CORS.")
 		r.Use(cors.Default())
 	}